rename \Rokka\Client\TemplateHelperCallbacksAbstract to \Rokka\Client\TemplateHelper\AbstractCallbacks
and \Rokka\Client\TemplateHelperDefaultCallbacks to \Rokka\Client\TemplateHelper\DefaultCallbacks
